Fixed bug where bullets would hit things both behind and in front of you.
git-svn-id: https://legesmotus.svn.sourceforge.net/svnroot/legesmotus/legesmotus/trunk@93 a5a5d5d7-22d1-44d1-b37a-67d3d99fcfd7

+++ common/math.hpp
@@ -30,7 +30,7 @@ static inline uint32_t toPow2(uint32_t num) {
 	return num;
 }
 
-vector<int> closest_point_on_line(int x1, int y1, int x2, int y2, int x3, int y3);
+vector<double> closest_point_on_line(int x1, int y1, int x2, int y2, int x3, int y3);
 
 static inline double dist_between_points(int x1, int y1, int x2, int y2) {
 	return sqrt(double((x2 - x1) * (x2 - x1) + (y2 - y1) * (y2 - y1)));
